Create ForkSystemCall
In the above program, we call the fork() system call to create a new process. If the fork() call returns 0, it means we are in the child process. The child process then calls the execlp() system call to execute the ls command using the /bin/ls program. This replaces the current process image with a new process image, which is the ls program. If the fork() call returns a positive value, it means we are in the parent process. The parent process waits for the child process to complete using the wait() system call. Finally, the program exits with a status code of 0.